backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
cryptography==41.0.1
distlib==0.3.6
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simple-apikey==0.1.2
djangorestframework-simplejwt==5.2.2
environ==1.0
filelock==3.12.0
pipenv==2023.5.19
platformdirs==3.5.1
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.23.0
virtualenv-clone==0.5.7
